requirements.txt
isort 
black
ruff
psycopg2
pytest
pytest-cov
flask
bootstrap-flask
flask_wtf
wtforms
pytest-mock
python-dotenv
SQLAlchemy